Refactoring mvvm hook standard convention
